Cambio de Integer a double

+++ pb2TrabajosPracticos/src/ar/edu/unlam/pb2/tpMonedero/Monedero.java
@@ -1,21 +1,21 @@
 package ar.edu.unlam.pb2.tpMonedero;
 
 public class Monedero {
-	private Integer dineroActual;
+	private double dineroActual;
 	
-	public Monedero (Integer dineroInicial) {
+	public Monedero (double dineroInicial) {
 		this.dineroActual = dineroInicial;
 	}
 	
-	public Integer consultarSaldo() {
+	public double consultarSaldo() {
 		return this.dineroActual;
 	}
 	
-	public void meterDinero(Integer dineroIngresado) {
+	public void meterDinero(double dineroIngresado) {
 		this.dineroActual += dineroIngresado;
 	}
 	
-	public void retirarDinero(Integer dineroRetirado) {
+	public void retirarDinero(double dineroRetirado) {
 		if(dineroRetirado < this.dineroActual) {
 			this.dineroActual -= dineroRetirado;
 		}
